Solve the problem of wrong geo-positioning in case of wrong reference to an axis of placement
Add the new method hasEntityById() to the class IfcFile that checks if the entity with the specified id exists.
Make the method TokenArgument::isNull() return true if the token is of type Token_IDENTIFIER and refers to non-existing entity.
After above changes, the following method:
  IfcGeom::Kernel::convert(const IfcSchema::IfcAxis2Placement3D* l, gp_Trsf& trsf)
will use the default direction gp_Dir(0,0,1) if the placement refers to non-existing main axis.